remove version from ProjectId
Most of the time we do not need the version together with the id.
E.g., in the CommandState.Waiting it is enough to know that we have a
dependency to the other project, the version is irrelevant. Moreover,
the version does not identify the project but rather a certain artifact.
Thus:
- move the `version` back to Project (name it currentVersion)
- introduce ProjectIdWithCurrentVersion, we still need the two together
  in the analysis
- adopt main, we do no longer require the version as argument