enable that a user can set a command to state Succeeded
We might consider a job as failed even though it is not for several
reasons:
- we got a timeout, after some time the jobs finishes
- we cannot extract a build number but the build started nonetheless
- we queued a Job but could not get build number (browser crashed)
- the remote releaser has a bug and thinks the job did not succeed but
  it did
- release job might fail (deploying fails) but releasing as such
  succeeded
- etc.

For such situations, it is handy if the user can set a command to state
succeeded so that she can continue with the release process.
There is also another use case where setting the state to succeeded is
handy, if we release a project (not with the dep-graph-releaser) but
still want to update the dependents. Then we set the root project's
release command to succeeded and start the process.