remove all toggle, is more confusing than helping
Users are confused that it is still active even though all commands
are deactivated. Some assume it stands for the release version. Others
are confused that it is not deactivated when all commands have succeeded
of the project but the release process as such failed.

I think there are only rare cases where you actually want to deactivate
all commands. Therefore it is simpler to remove this toggle so that we
can remove those confusions