do not use xpath narrowing during polling, due to chrome 404 bug
Chrome shows 404 requests always as error in the developer console
regardless if we catch the error or not. That is unfortunate but we can
circumvent this situation if we do not use xpath narrowing which returns
404 if it does not find a match. It might be a better idea to do it
on client side anyway because we should not load jenkins with this task.
Thus:
- remove xpath expression
- make sure we still match correctly