add excludeRegex to DependentsProjects console command
There are projects which are not submodules but do not reside in an own
repository. One shall be able to exclude such projects based on their
relative path. For instance, a project which resides not in a sub folder
but a sub sub folder (e.g. a/b/c) is not a repository.

- add mandatory parameter excludeRegex to DependentsProjects command
- adopt gradle task
- adopt jenkins.pom (add missing transform... arguments)