add relativePathToGitRepoRegex and co. to Config
We want to be able to generate an eclipse psf file from the GUI. For
this to work we need the relativePathToGitRepoRegex, the
relativePathToGitRepoReplacement and the
relativePathExcludeProjectRegex as we do for the console command
DependentProjects. Since this is most probably static per system we do
not want that the user has to provide it each time. Instead we add it
to the Config

- adopt Json
- adopt gradle task
- adopt jenkins.pom