give hint what `projects involved` means.
Some users are confused when they see number next to `Projects involved`
They think it should be lower since they see less projects.
We could split it up and show that submodules are included but I think
that would take up too much space at the bottom. But we can give the
user a tool tip and a question mark as mouse pointer indicating the
tool tip.