re-activate input fields of non successful projects
Currently we activate input fields if a process was successful. That is
only half clever. We should activate only those input fields which
belong to a project which has commands which are not yet in state
Succeeded. If a project has carried out all commands, then it does not
make sense that a user can change the version. However, if a project
has commands which failed or are deactivated then it makes sense.
Thus:
- activate input fields for project which have at least one
  non-successful/non-disabled command
- also reactivate config fields