add context menu for projects with git clone command
In case of a failure we might want to clone a specific project
and make some fixes. Providing a list of git clone commands with all
involved projects might not be what we want. Instead we could provide
a context menu per project with a context menu entry "git clone".

Moreover, cleanup:
- do not use releasePlan directly but only via ModifiedState
- get the surrounding project instead of MavenProject hack. For this
  to work we need to assign the project to the project div.