make when in recoverCommandStates exhaustive and fix parameterRegex
Also here it is better to be exhaustive so that a developer has to think
about it when introducing a new state. Moreover, we should not throw
an exception if state is ready (we have the TODO and it is documented
as known limitation).

The parameterRegex needs to be fixed since `<parameters>` can have
attributes.